[remove] 공통 컴포넌트 Header, HeaderTest 삭제 및 이름,경로 수정 #180
