--- src/components/common/HeaderTest/Header.module.css
@@ -1,98 +1,0 @@
-.header-wrap {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  border: 1px solid #dbdbdb;
-  padding: 8px 16px;
-}
-
-.btn-more,
-.btn-back {
-  background: none;
-}
-
-.test {
-  display: none;
-}
-
-/* HomeSearch 스타일 */
-.btn-search {
-  background: none;
-}
-
-.inp-search {
-  font-size: 18px;
-  width: inherit;
-  border: none;
-  color: #000000;
-  padding: 13px 16px;
-}
-.title {
-  font-weight: 500;
-  font-size: 18px;
-  line-height: 22px;
-  color: #000000;
-}
-
-/* UserSearch 스타일 */
-/* 다른건 크기조정안해도 유저서치안에 있는 이미지만 크기가 작아져서 조정함 */
-.img-back {
-  width: 22px;
-}
-.inp-userSearch {
-  width: inherit;
-  background-color: var(--bg-color);
-  border: 1px solid var(--main-color);
-  border-radius: 32px;
-  box-shadow: 0px 4px 4px var(--sub-color-2);
-  outline: none;
-  margin-left: 20px;
-  padding: 7px 19px;
-  font-size: 14px;
-}
-
-.inp-userSearch::placeholder {
-  color: #c4c4c4;
-}
-/* SaveButton 스타일 */
-.btn-save {
-  width: 90px;
-  height: 32px;
-  color: #fff;
-  font-size: 14px;
-  border-radius: 32px;
-  background-color: var(--sub-color-2);
-}
-
-.left {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  gap: 10px;
-}
-
-.chat-title {
-  font-size: 14px;
-}
-
-/* colorButton */
-.btn-save-color,
-.btn-upload-color {
-  width: 90px;
-  height: 32px;
-  color: #fff;
-  font-size: 14px;
-  border-radius: 32px;
-  background-color: var(--main-color);
-}
-
-.left {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  gap: 10px;
-}
-
-.chat-title {
-  font-size: 14px;
-}
